requirements.txt
requests
pytz
python-dotenv
ics
winotify